Add canUseBuilder helper method
